src/other/other.nosql/requirements.txt
boto3==1.34.143
cassandra-driver==3.29.1
pymongo==4.8.0
redis==5.0.7
docker==7.1.0